+++ Data/MRPData~Init.lua
@@ -1,7 +1,7 @@
 -- MRPData~Init.lua
 -- local _, MRPData = ...
 
-local VERSION = 10300
+local VERSION = 10400
 
 if MRPData_API and (MRPData_API.VERSION or 0) >= VERSION then return end