+++ MRP_Frames.lua
@@ -1,0 +1,5 @@
+-- MRP_Frames.lua
+-- local _, MRP = ...
+
+local Frames = {}
+MRP.Frames = Frames